Separated the two networks
Made new classes for the RBF network components because merge conflicts were gonna be hell at this point
